New concept for annotation in backend
